refactor(auth): minimal conversion of db/index to ts Because:
* We'd like more code to use TypeScript.

This commit:

* Minimally converts auth-server db/index to TS.
