+++ packages/fxa-auth-server/scripts/check-users.js
@@ -36,7 +36,8 @@ if (!program.input) {
 const log = require('../lib/log')({});
 const config = require('../config').config.getProperties();
 const Token = require('../lib/tokens')(log, config);
-const AuthDB = require('../lib/db')(config, log, Token);
+const { createDB } = require('../lib/db');
+const AuthDB = createDB(config, log, Token);
 const Password = require('../lib/crypto/password')(log, config);
 
 function sanitizeValue(v) {
