+++ packages/fxa-auth-server/scripts/delete-unverified-accounts.ts
@@ -9,7 +9,7 @@ import { processAccountDeletionInRange } from '../lib/routes/cloud-scheduler';
 
 import appConfig from '../config';
 import * as random from '../lib/crypto/random';
-import DB from '../lib/db';
+import { createDB } from '../lib/db';
 import { setupFirestore } from '../lib/firestore-db';
 import initLog from '../lib/log';
 import { CurrencyHelper } from '../lib/payments/currencies';
@@ -164,7 +164,7 @@ const init = async () => {
     { ...config.redis, ...config.redis.sessionTokens },
     log
   );
-  const db = DB(
+  const db = createDB(
     config,
     log,
     Token(log, config),
