+++ packages/fxa-auth-server/test/scripts/bulk-mailer.js
@@ -64,7 +64,8 @@ const account2Mock = createAccount(
   'es'
 );
 
-const DB = require('../../lib/db')(config, log, Token, UnblockCode);
+const { createDB } = require('../../lib/db');
+const DB = createDB(config, log, Token, UnblockCode);
 
 const execOptions = {
   cwd,
