+++ packages/fxa-auth-server/test/scripts/dump-users.js
@@ -57,7 +57,9 @@ const account2Mock = createAccount(
   crypto.randomBytes(16).toString('hex')
 );
 
-const DB = require('../../lib/db')(config, log, Token, UnblockCode);
+const { createDB } = require('../../lib/db');
+
+const DB = createDB(config, log, Token, UnblockCode);
 
 const execOptions = {
   cwd,
